requirements.txt
discord.py==2.1.0
PyNaCl==1.5.0
youtube_dl==2021.12.17
